libs/shared/data-access/phone-number-verification/src/lib/services/phone-number-verification.service.ts
import { Injectable } from '@nestjs/common';
import { ConfigService } from '@nestjs/config';

import { lastValueFrom, map, Observable } from 'rxjs';

import { OtpService } from '@oak/shared/data-access/arkesel';

import { SendCodeRequest } from '../interfaces';
import { VerifyCodeRequest } from '../interfaces';
import { Response } from '../interfaces';

@Injectable()
export class PhoneNumberVerificationService {
  constructor(
    private readonly config: ConfigService,
    private readonly otp: OtpService
  ) {}

  async sendCode(sendCodeRequest: SendCodeRequest): Promise<Response> {
    try {
      if (!sendCodeRequest.message.includes('%otp_code%')) {
        return {
          success: false,
          message: 'No OTP code variable found!',
        };
      }

      const generateOtpRequest = {
        number: sendCodeRequest.number,
        message: sendCodeRequest.message,
        expiry:
          sendCodeRequest.expiry ??
          this.config.get<number>('phoneNumberVerification.expiry'),
        senderId: this.config.get<string>('phoneNumberVerification.senderId'),
        type:
          sendCodeRequest.type ??
          this.config.get<'numeric' | 'alphanumeric'>(
            'phoneNumberVerification.type'
          ),
        medium:
          sendCodeRequest.medium ??
          this.config.get<'sms' | 'voice'>('phoneNumberVerification.medium'),
        length:
          sendCodeRequest.length ??
          this.config.get<number>('phoneNumberVerification.length'),
      };

      const generateOtp$: Observable<Response> = this.otp
        .generateOtp(generateOtpRequest)
        .pipe(
          map((res) => {
            if (res.success) {
              return {
                success: true,
                message: 'Code sent successfully',
                data: generateOtpRequest as SendCodeRequest,
              };
            }
            return {
              success: false,
              message: res.message,
            };
          })
        );

      return await lastValueFrom(generateOtp$);
    } catch (e) {
      return {
        success: false,
        message: 'Failed to send code',
      };
    }
  }

  async verifyCode(verifyCodeRequest: VerifyCodeRequest): Promise<Response> {
    try {
      const verifyOtp: Observable<Response> = this.otp
        .verifyOtp(verifyCodeRequest)
        .pipe(
          map((res) => {
            if (res.success) {
              return {
                success: true,
                message: 'Code has been verified',
                data: verifyCodeRequest,
              };
            }
            return {
              success: false,
              message: res.message,
            };
          })
        );

      return await lastValueFrom(verifyOtp);
    } catch (e) {
      return {
        success: false,
        message: 'Failed to verify code',
      };
    }
  }
}
